+++ tests/test_errors.py
@@ -1,3 +1,6 @@
+from scim2_models.exceptions import ExpiredCursorException
+from scim2_models.exceptions import InvalidCountException
+from scim2_models.exceptions import InvalidCursorException
 from scim2_models.exceptions import InvalidFilterException
 from scim2_models.exceptions import InvalidPathException
 from scim2_models.exceptions import InvalidSyntaxException
@@ -23,5 +26,8 @@ def test_predefined_errors():
         InvalidValueException(),
         InvalidVersionException(),
         SensitiveException(),
+        InvalidCursorException(),
+        ExpiredCursorException(),
+        InvalidCountException(),
     ):
         assert isinstance(exc.to_error(), Error)